Extract ExportMenu.java
